requirements.txt
fpdf==1.7.2
google_api_python_client==2.35.0
google_auth_oauthlib==0.4.6
mediapipe==0.8.9.1
numpy==1.21.4
opencv_python==4.5.4.60
pandas==1.3.5
Pillow==9.0.0
protobuf==3.19.3
